Fix type in vapply for total_days calculation
Changed the vapply return type from integer(1) to numeric(1) when summing n_days in compute_fsle_quartile_persistence. This ensures correct handling of numeric values for total_days.

+++ R/fsle_quartiles_global_hpc.R
@@ -174,7 +174,7 @@ compute_fsle_quartile_persistence <- function(fsle_dir,
     stop("[FSLE] No valid layers were processed. Check your input.")
   }
   
-  total_days <- sum(vapply(valid_res, function(x) x$n_days, integer(1)))
+  total_days <- sum(vapply(valid_res, function(x) x$n_days, numeric(1)))
   message("[FSLE] Total days processed across all files: ", total_days)
   
   # Sum partial rasters for each class
